Change and rejig some dependencies —
* move commons-configuration dependency from common module into the
  agent module because it's only ever needed on the agent, not on the
  server
* extract out an interface to break apart the
 `AgentAutoRegistrationProperties` dependency on commons-configuration
* upgrade `commons-lang` from 2.4 to 2.6 because commons-configuration
  needs it
* server module's websocket dependency should be `provides` because it's
  expected to be provided by the top level `go.jar` aka container
