+++ agent/pom.xml
@@ -65,6 +65,12 @@
             <version>${jetty.version}</version>
         </dependency>
 
+        <dependency>
+            <groupId>commons-configuration</groupId>
+            <artifactId>commons-configuration</artifactId>
+            <version>1.10</version>
+        </dependency>
+
         <!-- test dependencies -->
         <dependency>
             <groupId>com.thoughtworks.go</groupId>
